Migrate tests to pytest + add GA (#127)
